requirements.txt
# CloudSITLSIM Python Dependencies

# Web Framework
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
Flask-Migrate==4.0.5

# Database
SQLAlchemy==2.0.21
alembic==1.12.0

# Configuration
PyYAML==6.0.1
python-dotenv==1.0.0

# System utilities
psutil==5.9.5
pyserial==3.5

# MAVLink communication
pymavlink==2.4.37

# HTTP requests
requests==2.31.0

# Web server
gunicorn==21.2.0

# Development tools
pytest==7.4.2
pytest-cov==4.1.0
black==23.7.0
flake8==6.0.0

# Optional: For future telemetry features
# numpy==1.24.3
# matplotlib==3.7.2
# plotly==5.15.0
